+++ CHANGELOG.md
@@ -6,6 +6,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
 
+## Unreleased
+
+### Changed
+
+- SC-6474 in admin tables, edit column key is filtered out of column data object if school is external
+
+## [24.4.1] - 2020-09-08
+
+- SC-6701 - changed domain for ghost (mint-ec pages)
+
+## [24.4.0] - 2020-08-31
+
+- SC-5410 language dropdown in account page for selecting user language
+
+## [24.3.0] - 2020-08-25
+
+## [24.2.3] - 2020-08-19
+
+### Changed - 24.2.1
+
+- SC-6239 Changed default email domains in many different places from @schul-cloud.org to @hpi-schul-cloud.de
+- SC-6239 Changed links to blog to blog.hpi-schul-cloud.de
+
 ## [23.6.0] - 2020-07-21
 
 ### Changed